+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   `@lumx/core`:
     -   Moved `Uploader` stories and tests from `@lumx/react`
     -   Made `Uploader` `fileInputProps` type framework-agnostic
+    -   Moved `GenericBlock` stories and tests from `@lumx/react`
 
 ## [4.7.1][] - 2026-03-10